News Modal: Get rodal working
