+++ static/js/redux/ui/modals/Modal.tsx
@@ -1,6 +1,32 @@
+import React, { MouseEventHandler, ReactNode } from "react";
+// @ts-ignore
 import Rodal from "rodal";
-import "rodal/lib/rodal.css";
+import "./rodal.scss";
 
-const Modal = () => {};
+type RodalProps = {
+  children?: ReactNode;
+  width?: number;
+  height?: number;
+  measure?: string;
+  visible?: boolean;
+  showMask?: boolean;
+  closeOnEsc?: boolean;
+  closeMaskOnClick?: boolean;
+  showCloseButton?: boolean;
+  animation?: string;
+  enterAnimation?: string;
+  leaveAnimation?: string;
+  duration?: number;
+  className?: string;
+  customStyles?: { [key: string]: any };
+  customMaskStyles?: { [key: string]: any };
+  onClose: MouseEventHandler<HTMLSpanElement>;
+  onAnimationEnd?: () => never;
+};
+
+const Modal = (props: RodalProps) => {
+  const x = 0;
+  return <Rodal {...props}> Hello {x}</Rodal>;
+};
 
 export default Modal;
